align gdpr test with other extractor tests
