+++ tests/extractors/adblock_based_test.py
@@ -11,7 +11,7 @@ from features.addblock_based import (
     FanboySocialMedia,
 )
 from lib.tools import runtime
-from tests.integration.features_integration_test import mock_website_data
+from tests.extractors.conftest import mock_website_data
 
 
 @pytest.mark.asyncio
